backend/requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
alembic
annotated-types
anyio
appdirs
asgiref
asn1crypto
asttokens
attrs
auth0-python
Authlib
backoff
bcrypt
beautifulsoup4
blinker
build
cachetools
certifi
cffi
charset-normalizer
chroma-hnswlib
chromadb
click
cohere
colorama
coloredlogs
crewai
crewai-tools
cryptography
dataclasses-json
decorator
Deprecated
deprecation
distro
dnspython
docker
docstring_parser
docx2txt
durationpy
ecdsa
email_validator
embedchain
et_xmlfile
executing
fastapi
fastapi-cli
fastavro
filelock
flatbuffers
frozenlist
fsspec
gitdb
GitPython
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-aiplatform
google-cloud-bigquery
google-cloud-core
google-cloud-resource-manager
google-cloud-storage
google-crc32c
google-generativeai
google-resumable-media
googleapis-common-protos
gotrue
gptcache
greenlet
gritql
grpc-google-iam-v1
grpcio
grpcio-health-checking
grpcio-status
grpcio-tools
h11
h2
hpack
httpcore
httplib2
httptools
httpx
httpx-sse
huggingface-hub
humanfriendly
hyperframe
idna
ijson
importlib_metadata
importlib_resources
iniconfig
instructor
ipython
jedi
Jinja2
jiter
json5
json_repair
jsonpatch
jsonpickle
jsonpointer
jsonref
jsonschema
jsonschema-specifications
kubernetes
lancedb
langchain
langchain-cli
langchain-cohere
langchain-community
langchain-core
langchain-experimental
langchain-openai
langchain-text-splitters
langgraph
langgraph-checkpoint
langgraph-sdk
langserve
langsmith
linkup-sdk
litellm
Mako
markdown-it-py
MarkupSafe
marshmallow
matplotlib-inline
mdurl
mem0ai
mmh3
monotonic
mpmath
msgpack
multidict
mypy-extensions
networkx
nodeenv
numpy
oauthlib
onnxruntime
openai
openpyxl
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-fastapi
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-util-http
orjson
outcome
overrides
packaging
pandas
parameterized
parso
patronus
pdfminer.six
pdfplumber
pexpect
pillow
platformdirs
pluggy
portalocker
postgrest
posthog
prompt_toolkit
propcache
proto-plus
protobuf
ptyprocess
pure_eval
pyarrow
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic-settings
pydantic_core
Pygments
PyJWT
pylance
pymongo
pyOpenSSL
pyparsing
pypdf
pypdfium2
PyPika
pyproject_hooks
pyright
pysbd
PySocks
pytest
pytest-asyncio
python-dateutil
python-dotenv
python-jose
python-multipart
pytube
pytz
pyvis
PyYAML
qdrant-client
realtime
referencing
regex
requests
requests-oauthlib
requests-toolbelt
rich
rich-toolkit
rpds-py
rsa
schema
scrapegraph_py
selenium
serpapi
setuptools
shapely
shellingham
six
smmap
sniffio
snowflake-connector-python
sortedcontainers
soupsieve
spider-client
SQLAlchemy
sse-starlette
stack-data
starlette
storage3
StrEnum
supabase
supafunc
sympy
tabulate
tenacity
tiktoken
tokenizers
tomli
tomli_w
tomlkit
tqdm
traitlets
trio
trio-websocket
typer
types-requests
typing-inspect
typing_extensions
tzdata
uritemplate
urllib3
uv
uvicorn
uvloop
validators
watchfiles
wcwidth
weaviate-client
websocket-client
websockets
wrapt
wsproto
yarl
zipp